feat: grpc health check
Now only the provided connectionIds are accepted.
This is more secure.
The health check is needed because grpc is lazy and doesn't actually
establish a connection until it is absolutely needed. That means the the
first missile fired would have to wait roughly half a second until the
first control input arrives. That is way too slow.